bug: fix the mock client with a small refactor
